refactor: move fetching comments to container
